Create arduino_serial.ino
First cut at arduino code to take commands from the cRIO.
Right now it takes in various single character commands and turns
an LED on/off. This will be modified to interface with the
addressable rgb LED strip.

Send a for on, b for off, and c does on for one second and then off.