Modified and tuned in the auto mode command which will allow us to rotate to a commanded degree.
Had to add a rate limiter for increasing and decreasing, this may help
out for the drive straight function. Basically without the rate limiter,
the drivetrain motors would cause cRIO/access point restarts when
rotating. The command would go from 0,0 on the motor controllers
straight to something around 0.5, -0.5 and that was enough to cause
restarts.

We may need a rate limiter on the joystick commands into the drivetrain.
I was able to lose comm.s a few times using the sticks from stand still
to full forward. Not 100% repeatable.