Added auto and teleoperated command groups for getting the shooter and wheel up to position and speed.
Auto commands still needs some attention. The arm currently takes too
long to get to position for the command to work in 15 seconds though.

Tuned on robot.

Added PReset speeds and angles in RobotMap for different positions we
plan on shooting from on the field. Need real RPM values for the shooter
wheel.