Changes made at Gus's practice field 4/20/13. This code needs some cleanup.
Tested for the most part with new batteries around 13v
Only one encoder on the right side of the drivetrain.

One of the CIMs on the left (?) side of the drive train split open
towards the end of the day when we started seeing some weirdness with
the auto modes and communications dropping randomly... hopefully they're
related. Otherwise there's a seperate problem.